GH #1077 - Update the form after modifications.

+++ code/zato-web-admin/src/zato/admin/static/js/outgoing/redis/index.js
@@ -36,8 +36,8 @@ $.fn.zato.outgoing.redis.data_table.new_row = function(item, data, include_tr) {
         row += String.format("<tr id='tr_{0}' class='updated'>", item.id);
     }
 
-    var is_active = item.is_active == true;
-    var use_redis_sentinels = item.use_redis_sentinels == true;
+    var is_active = $.fn.zato.like_bool(item.is_active) == true;
+    var use_redis_sentinels = item.use_redis_sentinels  ? 'Yes' : 'No';
     var data_dicts_link = '/zato/outgoing/redis/data-dict/dictionary/?cluster=' + item.cluster_id;
     var remote_commands_link = '/zato/outgoing/redis/remote-command/?cluster=' + item.cluster_id;
 
@@ -56,13 +56,19 @@ $.fn.zato.outgoing.redis.data_table.new_row = function(item, data, include_tr) {
     row += String.format('<td>{0}</td>', use_redis_sentinels);
 
     // 3
-    row += String.format('<td>{0}</td>', item.redis_sentinels);
+    row += String.format('<td class="ignore">{0}</td>', item.redis_sentinels);
     row += String.format('<td>{0}</td>', String.format("<a href='javascript:$.fn.zato.data_table.change_password(\"{0}\")'>Change password</a>", item.id));
 
     // 4
-    row += String.format('<td>{0}</td>', item.redis_sentinels_master);
+    row += String.format('<td class="ignore">{0}</td>', item.redis_sentinels_master);
     row += String.format('<td>{0}</td>', String.format("<a href='javascript:$.fn.zato.outgoing.redis.edit(\"{0}\");'>Edit</a>", item.id));
-    row += String.format('<td>{0}</td>', String.format("<a href='javascript:$.fn.zato.outgoing.redis.delete_(\"{0}\");'>Delete</a>", item.id));
+
+    if(item.name != 'default') {
+        row += String.format('<td>{0}</td>', String.format("<a href='javascript:$.fn.zato.outgoing.redis.delete_(\"{0}\");'>Delete</a>", item.id));
+    }
+    else {
+        row += String.format('<td></td>');
+    }
 
     // 5
     row += String.format("<td class='ignore'>{0}</td>", item.use_redis_sentinels);
@@ -70,9 +76,9 @@ $.fn.zato.outgoing.redis.data_table.new_row = function(item, data, include_tr) {
     row += String.format("<td class='ignore'>{0}</td>", item.is_active);
 
     // 6
-    row += String.format('<td>{0}</td>', item.host);
-    row += String.format('<td>{0}</td>', item.port);
-    row += String.format('<td>{0}</td>', item.db);
+    row += String.format('<td class="ignore">{0}</td>', item.host);
+    row += String.format('<td class="ignore">{0}</td>', item.port);
+    row += String.format('<td class="ignore">{0}</td>', item.db);
 
     if(include_tr) {
         row += '</tr>';
